Add in try exception
When I was trying this, I entered a letter not a number.  It threw an exception.

So I thought I could add a try/except around the functioning code.  I tried a few ways, but I added it inside the while loop and now your code runs perfectly.

If there's an issue or a crash, it tells you there's an exception.